Rename to `button-has-type`
