+++ lib/rules/button-has-type.js
